Changes from review

+++ contracts/src/v0.8/functions/dev/1_0_0/FunctionsBilling.sol
@@ -20,7 +20,7 @@ abstract contract FunctionsBilling is Routable, IFunctionsBilling {
   using FunctionsResponse for FunctionsResponse.Commitment;
   using FunctionsResponse for FunctionsResponse.FulfillResult;
 
-  uint56 private constant REASONABLE_GAS_PRICE_CEILING = 1_000_000_000_000_000; // 1 million gwei
+  uint256 private constant REASONABLE_GAS_PRICE_CEILING = 1_000_000_000_000_000; // 1 million gwei
   // ================================================================
   // |                  Request Commitment state                    |
   // ================================================================
